requirements.txt
selenium
webdriver-manager
pytest
allure-pytest
pydantic
pydantic-settings
allure-python-commons
